refactor(session): move models to domain package
