+++ apps/session/internal/domain/quiz.go
@@ -1,4 +1,4 @@
-package models
+package domain
 
 type SelectionType string
