+++ apps/session/internal/domain/session.go
@@ -1,4 +1,4 @@
-package models
+package domain
 
 import "time"
