+++ apps/session/internal/service/session/interfaces.go
@@ -3,16 +3,16 @@ package session
 import (
 	"context"
 
-	"github.com/We-seek-D-p/quiz-platform/apps/session/internal/models"
+	"github.com/We-seek-D-p/quiz-platform/apps/session/internal/domain"
 )
 
 type ManagementClient interface {
-	GetBootstrap(ctx context.Context, sessionID string) (models.SessionBootstrap, error)
+	GetBootstrap(ctx context.Context, sessionID string) (domain.SessionBootstrap, error)
 }
 
 type RuntimeRepository interface {
-	Create(ctx context.Context, runtime models.SessionRuntime, quiz models.QuizSnapshot) error
-	Get(ctx context.Context, sessionID string) (models.SessionRuntime, error)
+	Create(ctx context.Context, runtime domain.SessionRuntime, quiz domain.QuizSnapshot) error
+	Get(ctx context.Context, sessionID string) (domain.SessionRuntime, error)
 	Delete(ctx context.Context, sessionID string) error
 }
